logger configured at module import
Added logger.py
