+++ handout/__init__.py
@@ -1,1 +1,5 @@
 from .handout import Handout
+
+from .logger import configure_logger
+
+configure_logger()
