+++ handout/tests/test_logger.py
@@ -3,5 +3,5 @@ import handout
 def test_get_logger_has_single_handler(): 
     import handout # Simulate multiple imports of a package.
     import handout
-    logger = handout.tools.get_logger()
+    logger = handout.logger.get_logger()
     assert len(logger.handlers) == 1
